✨ feat: 쇼핑카트에서 상품 하나 주문 POST 기능 구현 진행

+++ server/src/main/java/project2/SAYO/domain/shoppingCart/entity/ShoppingCartItem.java
@@ -31,7 +31,7 @@ public class ShoppingCartItem extends Auditable {
     private Item item;
 
     private String itemName;
-    private long itemPrice;
+    private int itemPrice;
     private String itemPicture;
     private int itemCount; // 담긴 상품의 수량
     private int itemTotalCount; // 장바구니에 담긴 상품의 총 수량
